backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
motor==3.4.0
pymongo==4.7.2
chromadb>=1.0.0
groq>=0.9.0
pdfplumber==0.11.0
PyMuPDF==1.24.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
reportlab==4.2.0
httpx==0.27.0
Pillow==10.3.0
python-pptx>=0.6.21
google-genai>=1.0.0
razorpay>=1.4.0
